refactor(api): move /api/certification-center-invitations/{id} route
to src/team
