+++ api/lib/application/certification-center-invitations/index.js
@@ -7,26 +7,6 @@ import { certificationCenterInvitationController } from './certification-center-
 const register = async function (server) {
   server.route([
     {
-      method: 'GET',
-      path: '/api/certification-center-invitations/{id}',
-      config: {
-        auth: false,
-        handler: certificationCenterInvitationController.getCertificationCenterInvitation,
-        validate: {
-          params: Joi.object({
-            id: identifiersType.certificationCenterInvitationId.required(),
-          }),
-          query: Joi.object({
-            code: Joi.string().required(),
-          }),
-        },
-        notes: [
-          "- Cette route permet de récupérer les détails d'une invitation selon un **id d'invitation** et un **code**\n",
-        ],
-        tags: ['api', 'invitations'],
-      },
-    },
-    {
       method: 'PATCH',
       path: '/api/certification-center-invitations/{certificationCenterInvitationId}',
       config: {
