+++ api/tests/team/acceptance/application/certification-center-invitation/certification-center-invitation.route.test.js
@@ -151,6 +151,42 @@ describe('Acceptance | Team | Application | Route | Certification Center Invitat
     });
   });
 
+  describe('GET /api/certification-center-invitations/{id}', function () {
+    it('returns the certification-center invitation and 200 status code', async function () {
+      // given
+      const certificationCenterId = databaseBuilder.factory.buildCertificationCenter({
+        name: 'Centre des Pixous',
+      }).id;
+      const certificationCenterInvitationId = databaseBuilder.factory.buildCertificationCenterInvitation({
+        certificationCenterId,
+        status: CertificationCenterInvitation.StatusType.PENDING,
+        code: 'ABCDEFGH01',
+      }).id;
+
+      await databaseBuilder.commit();
+
+      // when
+      const response = await server.inject({
+        method: 'GET',
+        url: `/api/certification-center-invitations/${certificationCenterInvitationId}?code=ABCDEFGH01`,
+      });
+
+      // then
+      expect(response.statusCode).to.equal(200);
+      expect(response.result).to.deep.equal({
+        data: {
+          type: 'certification-center-invitations',
+          id: certificationCenterInvitationId.toString(),
+          attributes: {
+            'certification-center-id': certificationCenterId,
+            'certification-center-name': 'Centre des Pixous',
+            status: CertificationCenterInvitation.StatusType.PENDING,
+          },
+        },
+      });
+    });
+  });
+
   describe('POST /api/certification-center-invitations/{id}/accept', function () {
     it('it returns an HTTP code 204', async function () {
       // given
